\name{frmpd-package}
\alias{frmpd-package}
\docType{package}
\title{Regression Analysis of Panel Fractional Responses}
\description{
Estimation of panel data regression models for fractional responses.
}
\details{
\tabular{ll}{
Package: \tab frmpd\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2015-02-18\cr
License: \tab GPL-2\cr
}
}
\author{Joaquim J.S. Ramalho <jsr@uevora.pt>}
\references{Ramalho, E.A. and J.J.S. Ramalho (2015),
"Exponential regression of panel data fractional response models", mimeo.
}