\name{frmpd-package}
\alias{frmpd-package}
\docType{package}
\title{Regression Analysis of Panel Fractional Responses}
\description{
Estimation of panel data regression models for fractional responses.
}
\details{
\tabular{ll}{
Package: \tab frmpd\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2016-08-03\cr
License: \tab GPL-2\cr
}
}
\author{Joaquim J.S. Ramalho <jsr@uevora.pt>}
\references{Ramalho, E.A. and J.J.S. Ramalho (2016), "Exponential regression of fractional-response fixed-efects models with an application to firm capital structure", mimeo.
}